security_system/requirements.txt
# Core LLM and AI
google-generativeai>=0.3.0

# Environment and HTTP
python-dotenv>=1.0.0
requests>=2.31.0

# Data processing
pandas>=2.0.0

# Dashboard UI
streamlit>=1.28.0

# Visualization (optional)
# plotly>=5.17.0

# Development and testing (optional)
# pytest>=7.4.0
# black>=23.0.0
# pylint>=3.0.0
